Msgbus: expect acknowledge byte from panel
Now an acknowledge byte (0xAC) is meant to be sent from the panel to
indicate receipt of a command and of extra data, only in the case where
the panel is not expected to send response data back (as then it'd be
redundant).

This is a form of flow control that ensures the IO board doesn't start
sending out additional data (after a command byte) before the panel is
ready to start receiving that data.